Move routes to main blueprint.
